dv.cc/iv.cc: When converting from an MTBDD, check that we don't write outside of the DoubleVector/IntegerVector
The conversion functions expect that the MTBDD is non-zero only for the reachable states
(as seen by the ODD). If this is not the case due to programmer error, writes outside
the allocated vector can happen and will lead to hard to debug phenomena.

The protection does a hard exit when it detects this problem.


git-svn-id: https://www.prismmodelchecker.org/svn/prism/prism/trunk@10549 bbc10eb1-c90d-0410-af57-cb519fbb1720